backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Daniel Maldonado <belialdaniel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.4"
Django = "^3.2.4"
python-dotenv = "^0.17.1"
GitPython = "^3.1.17"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.backend"
